+++ doc/conf.py
@@ -82,8 +82,6 @@ html_theme_path = [
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = []
-# html_static_path = ['_static']
-# htmlhelp_basename = "pygmshdoc"
 
 latex_elements = {
     "preamble": r"\usepackage[columns=1]{idxlayout}\makeindex",